relocate interfaces
