+++ src/pages/users/BulkUserUpload/BulkUserUpload.tsx
@@ -1,5 +1,5 @@
 import React, { ReactElement, useState } from "react";
-import { ImportUser, UploadedUser } from "../../../../Interfaces";
+import { ImportUser, UploadedUser } from "../../../Interfaces";
 import { NewUser } from "blaise-api-node-client";
 import UsersToUploadSummary from "./UsersToUploadSummary";
 import SelectFile from "./SelectFile";
